flake8: Add configuration file
This allows running flake8 on the code to check the formatting and some
common problems. The code is not currently compliant with flake8, so
flake8 is configured to ignore a lot of problems, but these could be
(gradually or selectively) fixed in the future.

This also tells flake8 to ignore the waf files, and tell it that `_` is
an additional builtin function (which is inserted into the builtin scope
by `hamster.lib.i18n.setup_i18n()` and `gettext.install()`).