NEWS: Use #nnn for all PR and issue references
This is more common notation, and when copying this file into github
(e.g. into the release notes or a commit message), it automatically
turns the references into the proper links. This does not work when
viewing the markdown file directly, unfortunately, that would require
adding explicit links which is harder to maintain and less readable in
plaintext.

This also removes the PR and issue prefixes, just the #nnn reference
looks cleaner (and a reader probably does not need to care whether it is
an issue or pullrequest, since they never have overlapping numbers).